requirements.txt
spotipy>=2.23.0
pandas>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.5.0
urllib3>=1.26.0
pytest>=7.3.1
pytest-mock>=3.10.0
logging>=0.4.9
